+++ src/main.c
@@ -6,14 +6,14 @@
 /*   By: dogwak <dogwak@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/08 18:21:43 by dogwak            #+#    #+#             */
-/*   Updated: 2024/08/09 18:45:46 by dogwak           ###   ########.fr       */
+/*   Updated: 2024/08/09 20:32:07 by dogwak           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "raytracer.h"
 #include <stdlib.h>
 #include <unistd.h>
 #include "./libft/libft.h"
+#include "./raytracer/raytracer.h"
 
 int	main(int argc, char *argv[])
 {
