+++ src/ray/ray.h
@@ -6,14 +6,14 @@
 /*   By: dogwak <dogwak@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/05 11:01:54 by dogwak            #+#    #+#             */
-/*   Updated: 2024/08/05 12:10:53 by dogwak           ###   ########.fr       */
+/*   Updated: 2024/08/09 20:25:07 by dogwak           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef RAY_H
 # define RAY_H
 
-# include "./ft_graphics/ft_math/ft_math.h"
+# include "../ft_graphics/ft_math/ft_math.h"
 
 typedef struct s_ray
 {
