+++ src/raytracer/init_resource.c
