+++ src/raytracer/raytracer.h
@@ -6,7 +6,7 @@
 /*   By: dogwak <dogwak@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/05 11:21:17 by dogwak            #+#    #+#             */
-/*   Updated: 2024/08/09 15:26:45 by dogwak           ###   ########.fr       */
+/*   Updated: 2024/08/09 20:25:24 by dogwak           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,8 @@
 # define SCREEN_WIDTH 1920
 # define SCREEN_HEIGHT 1080
 
-# include "./ft_graphics/ft_graphics.h"
-# include "./ft_vector/ft_vector.h"
+# include "../ft_graphics/ft_graphics.h"
+# include "../ft_vector/ft_vector.h"
 
 /*
 	ftmlx	: mlx object
