+++ src/raytracer/render_image.c
