+++ src/raytracer/utility.c
